Add files via upload
I have a form I'm building that uploads the form content into a database (MySQL) and a separate page to display the contents of the database. So far everything works great. I just needed the users to be able to upload an image file along with the form and I need the image itself and a few more fields that I may add to display on the page with the database contents. I have created the fields and the image uploader but when I run my program, it does not display the added fields and photo. I have since created new columns for the new fields in my database already.

How do I accomplish this by modifying the existing code?